pyproject.toml
[tool.poetry]
name = "Altinn Studio Assistent"
version = "0.1.0"
description = ""
authors = ["Open Source contributors"]
license = ""
readme = "README.md"

[tool.poetry.scripts]
start = "main:app"

[tool.poetry.dependencies]
python = "^3.9"
ctransformers = "^0.2.26"
faiss-cpu = "^1.7.4"
fastapi = "^0.96.0"
ipykernel = "^6.23.1"
langchain = "^0.0.225"
pypdf = "3.8.1"
python-box = "7.0.1"
python-dotenv = "1.0.0"
sentence-transformers = "2.2.2"
uvicorn = "^0.22.0"
torch = ">=2.0.0, !=2.0.1"
unstructured = "^0.10.10"
markdown = "^3.4.4"
openai = "^0.27.10"
slack-bolt = "^1.18.0"
tiktoken = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
